+++ egui_commonmark_backend/src/lib.rs
@@ -11,6 +11,9 @@ pub mod misc;
 #[doc(hidden)]
 pub mod pulldown;
 
+#[cfg(feature = "embedded_image")]
+mod data_url_loader;
+
 // For ease of use in proc macros
 #[doc(hidden)]
 pub use {